added some features

+++ friends/friendship/views.py
@@ -7,23 +7,63 @@ from django.contrib.auth.models import User
 from django.contrib import messages
 from django.shortcuts import HttpResponse
 
-def index(request):
+def index(request): # Главная страница
     return render(request, 'friendship/base.html')
 
+@login_required
+def outbox(request): # Исходящие запросы в друзья
+    sent_requests = FriendRequest.objects.filter(from_user=request.user, accepted=0)
+
+    return render(request, 'friendship/sent_friend_requests.html', {'sent_requests': sent_requests})
+
+@login_required
 def inbox(request):
-    return HttpResponse("syka blyat nahui in")
+    # получаем все заявки в друзья,полученные текущим пользователем
+    friend_requests_received = FriendRequest.objects.filter(to_user=request.user, accepted=0)
 
-def outbox(request):
-    return HttpResponse("syka blyat nahui out")
+    return render(request, 'friendship/received_friend_requests.html', {'incoming_requests': friend_requests_received})
 
-def register(request):
+@login_required
+def cancel_friend_request(request, friend_request_id): # Отмена запроса
+    friend_request = get_object_or_404(FriendRequest, id=friend_request_id)
+
+    # проверяем, является ли пользователь отправителем запроса
+    if friend_request.from_user != request.user:
+        return redirect('friendship/outbox')
+
+    friend_request.delete()
+
+    return redirect('/friendship/outbox')
+
+@login_required
+def open_status_field(request): # Создание формы для получения статуса
+    if request.method == 'POST':
+        user_to_check = request.POST.get("username")
+        print(request)
+        return redirect(f'/friendship/get_status/{user_to_check}')
+    else:
+        return render(request, 'friendship/status_field.html')
+
+@login_required
+def get_status(request, username):
+    # Находим пользователя по имени пользователя
+    user_to_ckeck = get_object_or_404(User, username=username)
+
+    # Получаем статус дружбы между текущим пользователем и профилем пользователя
+    friendship_status = _get_friendship_status(request.user, user_to_ckeck)
+
+    context = {
+        'status': friendship_status,
+        'checked_user': user_to_ckeck,
+    }
+
+    return render(request, 'friendship/status.html', context)
+
+def register(request): # Регистрация нового пользователя
     if request.method == 'POST':
         form = UserCreationForm(request.POST)
-        print("1")
         if form.is_valid():
-            print("2")
             form.save()
-            print("3")
             return redirect('/friendship/login')
         else:
             print(form.errors)
@@ -31,7 +71,7 @@ def register(request):
         form = UserCreationForm()
     return render(request, 'friendship/registration.html', {'form': form})
 
-def login_view(request):
+def login_view(request): # Авторизация пользователя
     if request.method == 'POST':
         form = AuthenticationForm(request, request.POST)
         if form.is_valid():
@@ -46,19 +86,7 @@ def login_view(request):
     return render(request, 'friendship/login.html', {'form': form})
 
 @login_required
-def friend_requests(request):
-    # получаем все заявки в друзья, отправленные или полученные текущим пользователем
-    friend_requests_sent = request.user.friend_requests_sent.all()
-    friend_requests_received = request.user.friend_requests_received.all()
-
-    context = {
-        'sent': friend_requests_sent,
-        'received': friend_requests_received,
-    }
-    return render(request, 'friendship/friend_requests.html', context)
-
-@login_required
-def create_request(request):
+def create_request(request): # Создание формы для запроса в друзья
     if request.method == 'POST':
         recipient_username = request.POST.get("recipient_username")
         print(request)
@@ -67,31 +95,82 @@ def create_request(request):
         return render(request, 'friendship/request_creation.html')
 
 @login_required
-def send_friend_request(request, recipient_username):
+def send_friend_request(request, recipient_username): # Создание запроса в друзья
     recipient = get_object_or_404(User, username=recipient_username)
 
     # проверяем, существует ли уже запрос на дружбу между пользователями
     if FriendRequest.objects.filter(from_user=request.user, to_user=recipient).exists():
         messages.error(request, 'A friend request has already been sent to this user.')
-        return redirect('user_profile', username=recipient_username)
+        return redirect('/friendship/outbox')
+    
+    if FriendRequest.objects.filter(from_user=recipient, to_user=request.user).exists():
+        received_request = FriendRequest.objects.get(from_user=recipient, to_user=request.user)
+        received_request.accepted = True
+        received_request.save()
+        return redirect('/friendship/friends')
 
     # создаем объект FriendRequest и сохраняем его в базе данных
     friend_request = FriendRequest(from_user=request.user, to_user=recipient)
     friend_request.save()
     messages.success(request, 'Friend request sent.')
-    return redirect('user_profile', username=recipient_username)
+    return redirect('/friendship/outbox')
 
 @login_required
-def accept_friend_request(request, pk):
-    friend_request = get_object_or_404(FriendRequest, pk=pk, to_user=request.user)
+def accept_friend_request(request, friend_request_id):
+    friend_request = get_object_or_404(FriendRequest, id=friend_request_id)
+
+    # проверяем, является ли пользователь получателем запроса
+    if friend_request.to_user != request.user:
+        return redirect('/friendship/inbox')
+
+    # соглашаемся на дружбу
     friend_request.accepted = True
     friend_request.save()
-    messages.success(request, 'Friend request accepted.')
-    return redirect('friend_requests')
+
+    return redirect('/friendship/inbox')
 
 @login_required
-def reject_friend_request(request, pk):
-    friend_request = get_object_or_404(FriendRequest, pk=pk, to_user=request.user)
+def reject_friend_request(request, friend_request_id):
+    friend_request = get_object_or_404(FriendRequest, id=friend_request_id)
+
+    # проверяем, является ли пользователь получателем запроса
+    if friend_request.to_user != request.user:
+        return redirect('/friendship/inbox')
+
     friend_request.delete()
-    messages.success(request, 'Friend request rejected.')
-    return redirect('friend_requests')
+
+    return redirect('/friendship/inbox')
+
+def friend_list(request):
+    # Находим всех друзей пользователя
+    friends = FriendRequest.objects.filter(from_user=request.user, accepted=True) | FriendRequest.objects.filter(to_user=request.user, accepted=True)
+
+    context = {
+        'friends': friends
+    }
+    return render(request, 'friendship/friend_list.html', context)
+
+def _get_friendship_status(current_user, other_user):
+    """Проверяет статус дружбы между двумя пользователями"""
+    friend_request_sent = FriendRequest.objects.filter(
+        from_user=current_user,
+        to_user=other_user
+    ).exists()
+    
+    friend_request_received = FriendRequest.objects.filter(
+        from_user=other_user,
+        to_user=current_user
+    ).exists()
+
+
+    
+    if friend_request_sent and friend_request_received:
+        friendship_status = 'Друзья'
+    elif friend_request_sent:
+        friendship_status = 'Заявка отправлена'
+    elif friend_request_received:
+        friendship_status = 'Заявка получена'
+    else:
+        friendship_status = 'Не друзья'
+    
+    return friendship_status
